Move tests to the matching modules
