+++ tests/unit/source_code/test_graph.py
@@ -1,4 +1,3 @@
-from collections.abc import Iterable
 from pathlib import Path
 
 import pytest
@@ -11,10 +10,7 @@ from databricks.labs.ucx.source_code.graph import (
     DependencyProblem,
     InheritedContext,
 )
-from databricks.labs.ucx.source_code.linters.graph_walkers import DependencyGraphWalker
 from databricks.labs.ucx.source_code.notebooks.loaders import NotebookLoader
-from databricks.labs.ucx.source_code.path_lookup import PathLookup
-from databricks.labs.ucx.source_code.python.python_ast import Tree
 
 
 def test_dependency_graph_registers_library_from_egg(mock_path_lookup, simple_dependency_resolver) -> None:
@@ -193,31 +189,6 @@ def test_graph_builds_inherited_context(mock_path_lookup, simple_dependency_reso
     assert inference_context.tree is not None
     assert inference_context.tree.has_global("some_table_name")
     assert not inference_context.tree.has_global("other_table_name")
-
-
-def test_graph_walker_captures_lineage(mock_path_lookup, simple_dependency_resolver) -> None:
-    grand_parent = mock_path_lookup.cwd / "functional/grand_parent_that_magic_runs_parent_that_magic_runs_child.py"
-    child = mock_path_lookup.cwd / "functional/_child_that_uses_value_from_parent.py"
-    root_dependency = Dependency(NotebookLoader(), grand_parent)
-    root_graph = DependencyGraph(
-        root_dependency, None, simple_dependency_resolver, mock_path_lookup, CurrentSessionState()
-    )
-    container = root_dependency.load(mock_path_lookup)
-    assert container is not None
-    container.build_dependency_graph(root_graph)
-
-    class _TestWalker(DependencyGraphWalker):
-        def _process_dependency(
-            self, dependency: Dependency, path_lookup: PathLookup, inherited_tree: Tree | None
-        ) -> Iterable[None]:
-            if dependency.path.as_posix().endswith(grand_parent.as_posix()):
-                assert len(self._lineage) == 1
-            if dependency.path.as_posix().endswith(child.as_posix()):
-                assert len(self._lineage) == 3  # there's a parent between grand_parent and child
-            return []
-
-    walker = _TestWalker(root_graph, set(), mock_path_lookup)
-    _ = list(_ for _ in walker)
 
 
 def test_dependency_problem_has_path_missing_by_default() -> None:
