+++ tests/unit/source_code/test_jobs.py
@@ -20,7 +20,6 @@ from databricks.sdk.errors import NotFound
 from databricks.sdk.service import compute, jobs, pipelines
 from databricks.sdk.service.workspace import ExportFormat, ObjectInfo, Language
 
-from databricks.labs.ucx.source_code.linters.context import LinterContext
 from databricks.labs.ucx.source_code.linters.files import FileLoader, ImportFileResolver
 from databricks.labs.ucx.source_code.graph import (
     Dependency,
@@ -28,7 +27,6 @@ from databricks.labs.ucx.source_code.graph import (
     DependencyResolver,
 )
 from databricks.labs.ucx.source_code.jobs import JobProblem, WorkflowLinter, WorkflowTaskContainer
-from databricks.labs.ucx.source_code.linters.graph_walkers import LintingWalker
 from databricks.labs.ucx.source_code.notebooks.loaders import NotebookResolver, NotebookLoader
 from databricks.labs.ucx.source_code.used_table import UsedTablesCrawler
 
@@ -515,21 +513,6 @@ def test_xxx(graph) -> None:
     assert workflow_task_container.spark_conf == {"spark.databricks.cluster.profile": "singleNode"}
 
     ws.assert_not_called()
-
-
-def test_linting_walker_populates_paths(dependency_resolver, mock_path_lookup, migration_index) -> None:
-    path = mock_path_lookup.resolve(Path("functional/values_across_notebooks_dbutils_notebook_run.py"))
-    root = Dependency(NotebookLoader(), path)
-    xgraph = DependencyGraph(root, None, dependency_resolver, mock_path_lookup, CurrentSessionState())
-    current_session = CurrentSessionState()
-    walker = LintingWalker(
-        xgraph, set(), mock_path_lookup, "key", current_session, lambda: LinterContext(migration_index, current_session)
-    )
-    advices = 0
-    for advice in walker:
-        advices += 1
-        assert "UNKNOWN" not in advice.path.as_posix()
-    assert advices
 
 
 def test_workflow_linter_refresh_report(dependency_resolver, mock_path_lookup, migration_index) -> None:
